backend/src/services/cognito_service.py
# For JWT verification, use python-jose or cognitojwt
from jose import jwt
import requests
from core.config import COGNITO_USER_POOL_ID, COGNITO_CLIENT_ID, AWS_REGION

# Add JWT verification logic here (see AWS docs for details)